@@ -1,12 +1,236 @@
-var React = require("react"),
-  ReactDOM = require("react-dom"),
+import React, { useState } from "react";
+var ReactDOM = require("react-dom"),
   d3 = require("d3"),
   _ = require("underscore");
 
-import { Tree } from "./components/tree.jsx";
+import ReactTree, {
+  ModelsPartitionsList,
+  SettingsItem
+} from "./components/react-tree.jsx";
 import { DatamonkeyTable, DatamonkeyModelTable } from "./components/tables.jsx";
 import { DatamonkeySiteGraph } from "./components/graphs.jsx";
 import { ResultsPage } from "./components/results_page.jsx";
+
+function MEMETree(props) {
+  const { data } = props;
+  if (!data) return null;
+  const ebf_data = _.object(
+      _.keys(data["branch attributes"]["attributes"])
+        .filter(datum => datum.slice(0, 3) == "EBF")
+        .map(datum => [+datum.split(" ")[2], +datum[datum.length - 2]])
+    ),
+    ebf_range = d3.extent(_.keys(ebf_data).map(d => +d));
+  const [partition, setPartition] = useState(0);
+  const [model, setModel] = useState("Global MG94xREV");
+  const [ebf, setEbf] = useState(null);
+  const [message, setMessage] = useState(null);
+  const [threshold, setThreshold] = useState(1);
+  const newick = data.input.trees[String(partition)];
+  function accessor(node) {
+    const name = node.data.name,
+      attributes = data["branch attributes"][String(partition)],
+      branch_length = attributes[name][model];
+    return branch_length;
+  }
+
+  var branchStyler, labelStyler;
+  if (ebf != null && ebf_data[ebf]) {
+    const current_partition = data["branch attributes"][String(partition)],
+      ebf_string = `EBF site ${ebf} (partition ${partition + 1})`;
+    labelStyler = node => {
+      const ebf_value = current_partition[node.name][ebf_string],
+        above_threshold = ebf_value > threshold,
+        fill = above_threshold ? "red" : "lightgrey",
+        opacity = above_threshold ? 1 : 0.8;
+      return {
+        fill: fill,
+        opacity: opacity
+      };
+    };
+    branchStyler = node => {
+      const ebf_value = current_partition[node.name][ebf_string],
+        above_threshold = ebf_value > threshold,
+        stroke = above_threshold ? "red" : "lightgrey",
+        opacity = above_threshold ? 1 : 0.8;
+      return {
+        stroke: stroke,
+        opacity: opacity
+      };
+    };
+  } else {
+    branchStyler = null;
+    labelStyler = null;
+  }
+  return (
+    <ReactTree
+      newick={newick}
+      accessor={accessor}
+      number_of_sequences={props.data.input["number of sequences"]}
+      paddingLeft={ebf == null ? 20 : 100}
+      popover={
+        "<li>Use the options menu to toggle the different site partitions.</li>"
+      }
+      branchStyler={branchStyler}
+      labelStyler={labelStyler}
+      includeBLAxis
+      options={
+        <ModelsPartitionsList
+          number_of_partitions={_.keys(data.input.trees).length}
+          partition={partition}
+          setPartition={setPartition}
+          model={model}
+          setModel={setModel}
+        />
+      }
+      settings={
+        <SettingsItem
+          onClick={() => {
+            setEbf(ebf == null ? ebf_range[0] : null);
+          }}
+        >
+          {ebf == null ? "Show" : "Hide"} branch-wise EBF values
+        </SettingsItem>
+      }
+      between={
+        ebf != null ? (
+          <div
+            style={{
+              display: "flex",
+              justifyContent: "space-between",
+              marginLeft: 20,
+              marginTop: 10,
+              marginBottom: 10,
+              height: 50
+            }}
+          >
+            <span>
+              Site to display:
+              <input
+                value={ebf}
+                size={5}
+                onInput={e => {
+                  const site = +e.target.value,
+                    partition = ebf_data[site];
+                  if (isNaN(site) && site != "") {
+                    setMessage({
+                      type: "error",
+                      content: "Please enter a valid input (positive integer)."
+                    });
+                    setEbf("");
+                  } else if (site == "") {
+                    setEbf("");
+                  } else if (site > ebf_range[1]) {
+                    setMessage({
+                      type: "warning",
+                      content: `Maximal site with EBf info is ${ebf_range[1]}.`
+                    });
+                    setEbf(ebf_range[1]);
+                  } else if (site < 1) {
+                    setMessage({
+                      type: "error",
+                      content: "Please enter a positive number."
+                    });
+                    setEbf(1);
+                  } else if (!ebf_data[site]) {
+                    setMessage({
+                      type: "warning",
+                      content: "No variation at this site, so no EBFs."
+                    });
+                    setEbf(+site);
+                  } else {
+                    setMessage(null);
+                    setEbf(+site);
+                    setPartition(partition - 1);
+                  }
+                }}
+              />
+              <i
+                className="fa fa-plus-circle"
+                aria-hidden="true"
+                onClick={e => {
+                  const new_ebf = ebf + 1;
+                  if (new_ebf < ebf_range[1] && ebf_data[new_ebf]) {
+                    setMessage(null);
+                    setEbf(new_ebf);
+                  } else if (new_ebf < ebf_range[1]) {
+                    setMessage({
+                      type: "warning",
+                      content: "No variation at this site, so no EBFs."
+                    });
+                    setEbf(new_ebf);
+                  } else {
+                    setMessage({
+                      type: "warning",
+                      content: `Maximal site with EBf info is ${ebf_range[1]}.`
+                    });
+                  }
+                }}
+              />
+              <i
+                className="fa fa-minus-circle"
+                aria-hidden="true"
+                onClick={e => {
+                  const new_ebf = ebf - 1;
+                  if (new_ebf > 0 && ebf_data[new_ebf]) {
+                    setMessage(null);
+                    setEbf(new_ebf);
+                  } else if (new_ebf > 0) {
+                    setMessage({
+                      type: "warning",
+                      content: "No variation at this site, so no EBFs."
+                    });
+                    setEbf(new_ebf);
+                  } else {
+                    setMessage({
+                      type: "warning",
+                      content: "Please enter a positive number."
+                    });
+                  }
+                }}
+              />
+            </span>
+
+            <span>
+              <span style={{ width: 300, display: "inline-block" }}>
+                <span>EBF threshold:</span>
+                <input
+                  type="range"
+                  style={{ width: 180 }}
+                  min="0"
+                  max="10"
+                  step=".1"
+                  value={threshold}
+                  onChange={e => {
+                    setThreshold(e.target.value);
+                  }}
+                />
+              </span>
+              <span
+                style={{ width: 30, marginLeft: 10, display: "inline-block" }}
+              >
+                {d3.format("1.1f")(threshold)}
+              </span>
+            </span>
+
+            <span>
+              {message ? (
+                <div
+                  className={
+                    "alert alert-" +
+                    (message.type == "error" ? "danger" : "warning")
+                  }
+                  role="alert"
+                >
+                  <b> {message.type.toUpperCase()}: </b> {message.content}
+                </div>
+              ) : null}
+            </span>
+          </div>
+        ) : null
+      }
+    />
+  );
+}
 
 function MEMESummary(props) {
   var number_of_sites = 0;
@@ -272,13 +496,7 @@ class MEMEContents extends React.Component {
 
         <div className="row">
           <div id="tree-tab" className="col-md-12">
-            <Tree
-              models={models}
-              json={self.state.data}
-              settings={tree_settings}
-              method={"meme"}
-              multitree
-            />
+            <MEMETree data={self.state.data} />
           </div>
         </div>
 
